自解析打包解包逻辑

+++ protolua/message.cpp
@@ -14,7 +14,7 @@ bool message_pack(lua_State* L, int start, int end, char* output, size_t* size)
     size_t proto_len = 0;
     const char* proto = lua_tolstring(L, start, &proto_len);
     int head_len = encode_varint(output, left_size, proto_len);
-    if (head_len < 0 || left_size < head_len + proto_len)
+    if (head_len < 0 || left_size < head_len + (int)proto_len)
     {
         proto_error("message_pack buffer limit, proto=%s", proto);
         return false;
@@ -36,7 +36,7 @@ bool message_pack(lua_State* L, int start, int end, char* output, size_t* size)
 // [proto, filed1, field2, ...]
 bool message_unpack(lua_State* L, const char* input, size_t size)
 {
-    int proto_len = 0;
+    uint64_t proto_len = 0;
     int head_len = decode_varint(&proto_len, input, size);
     if (head_len <= 0)
     {
@@ -44,17 +44,17 @@ bool message_unpack(lua_State* L, const char* input, size_t size)
         return false;
     }
 
-    if (size < head_len + proto_len)
+    if ((int)size < head_len + proto_len)
     {
         proto_error("message_unpack buffer not enough proto, proto_len=%d", proto_len);
         return false;
     }
 
     int top = lua_gettop(L);
-    std::string proto(input + head_len, proto_len);
+    std::string proto(input + head_len, (size_t)proto_len);
     lua_pushlstring(L, proto.data(), proto.size());
-    input += head_len + proto_len;
-    size -= head_len + proto_len;
+    input += head_len + (int)proto_len;
+    size -= head_len + (int)proto_len;
 
     if (!proto_unpack(proto.c_str(), L, input, size))
     {
